str.emmGrid {emmeans}	R Documentation

Miscellaneous methods for emmGrid objects

Description

Miscellaneous methods for emmGrid objects

Usage

## S3 method for class 'emmGrid'
str(object, ...)

## S3 method for class 'emmGrid'
print(x, ..., export = FALSE)

## S3 method for class 'emmGrid'
vcov(object, ...)

Arguments

object
An emmGrid object

...
(required but not used)

x
An emmGrid object

export
Logical value. If FALSE, the object is printed. If TRUE, a list is invisibly returned, which contains character elements named summary and annotations that may be saved or displayed as the user sees fit. summary is a character matrix (or list of such matrices, if a by variable is in effect). annotations is a character vector of the annotations that would have been printed below the summary or summaries.

Value

The vcov method returns a symmetric matrix of variances and covariances for predict.emmGrid(object, type = "lp")

[Package emmeans version 1.8.2 Index]