cld {emmeans}	R Documentation

Temporary continued 'cld' support

Description

This generic is temporarily provided to keep downstream packages from breaking. It will be removed in future versions to avoid masking issues with multcomp::cld. Please use CLD instead when working with emmMeans objects.

Usage

cld(object, ...)

Arguments

object
an emmGrid object

...
other arguments passed to CLD

[Package emmeans version 1.3.2 Index]