emm_list {emmeans}	R Documentation

The emm_list class

Description

An emm_list object is simply a list of emmGrid objects. Such a list is returned, for example, by emmeans with a two-sided formula or a list as its specs argument.

Details

Methods for emm_list objects include summary, CLD, coef, confint, contrast, pairs, plot, print, and test. These are all the same as those methods for emmGrid objects, with an additional which argument (integer) to specify which members of the list to use. The default is which = seq_along(object); i.e., the method is applied to every member of the emm_list object. The exception is plot, where only the which[1]th element is plotted.

As an example, to summarize a single member – say the second one – of an emm_list, one may use summary(object, which = 2), but it is probably preferable to directly summarize it using summary(object[[2]]).

[Package emmeans version 1.4.5 Index]