emm_list {emmeans}	R Documentation

The emm_list class

Description

An emm_list object is simply a list of emmGrid objects. Such a list is returned, for example, by emmeans with a two-sided formula or a list as its specs argument. Several methods for this class are provided, as detailed below.

Usage

## S3 method for class 'emm_list'
contrast(object, ..., which = I_bet(1))

## S3 method for class 'emm_list'
pairs(x, ..., which = I_bet(1))

## S3 method for class 'emm_list'
test(object, ..., which = seq_along(object))

## S3 method for class 'emm_list'
confint(object, ..., which = seq_along(object))

## S3 method for class 'emm_list'
plot(x, ..., which = I_bet(1))

## S3 method for class 'emm_list'
coef(object, ..., which = I_bet(1))

## S3 method for class 'emm_list'
str(object, ...)

## S3 method for class 'emm_list'
summary(object, ..., which = seq_along(object))

## S3 method for class 'emm_list'
print(x, ...)

## S3 method for class 'emm_list'
as.data.frame(x, ...)

I_bet(which)

Arguments

object, x
an object of class emm_list

...
additional arguments passed to corresponding emmGrid method

which
integer vector specifying which elements to select.

Value

a list of objects returned by the corresponding emmGrid method (thus, often, another emm_list object). However, if which has length 1, the one result is not wrapped in a list.

The as.data.frame method returns a single data frame via as.data.frame(rbind(x)). See also rbind.emm_list and as.data.frame.emmGrid

Note

The plot method uses only the first element of which; the others are ignored.

No export option is provided for printing an emm_list (see print.emmGrid). If you wish to export these objects, you must do so separately for each element in the list.

I_bet() provides a default value for which noisily. It is used in cases where users most likely intended to call a method for an emmGrid object rather than an emm_list object. An informative message is displayed and then the whichth element is returned.

[Package emmeans version 1.8.2 Index]