The emm_list class

An emm_list object is simply a list of emmGrid objects. Such a list is returned, for example, by emmeans with a two-sided formula or a list as its specs argument. Several methods for this class are provided, as detailed below. Typically, these methods just quietly do the same thing as their emmGrid methods, using the first element of the list. You can specify which to select a different element, or just run the corresponding emmGrid method on object[[k]].

Usage

# S3 method for emm_list
contrast(object, ..., which = 1)

# S3 method for emm_list
pairs(x, ..., which = 1)

# S3 method for emm_list
test(object, ..., which = seq_along(object))

# S3 method for emm_list
confint(object, ..., which = seq_along(object))

# S3 method for emm_list
plot(x, ..., which = 1)

# S3 method for emm_list
coef(object, ..., which = 2)

# S3 method for emm_list
str(object, ...)

# S3 method for emm_list
summary(object, ..., which = seq_along(object))

# S3 method for emm_list
print(x, ...)

# S3 method for emm_list
as.data.frame(x, ...)

# S3 method for summary_eml
as.data.frame(x, row.names = NULL, optional = FALSE,
  which, ...)

Arguments

object, x

an object of class emm_list

...

additional arguments passed to corresponding emmGrid method

which

integer vector specifying which elements to select.

row.names, optional

Required arguments of as.data.frame, ignored

Value

a list of objects returned by the corresponding emmGrid

method (thus, often, another emm_list object). However, if

which has length 1, the one result is not wrapped in a list.

summary.emm_list returns an object of class summary_eml, which is a list of summary_emm

objects.

The as.data.frame methods return a single data frame via as.data.frame(rbind(x)). See also rbind.emm_list and as.data.frame.emmGrid

Note

The plot method uses only the first element of which; the others are ignored.

No export option is provided for printing an emm_list (see print.emmGrid). If you wish to export these objects, you must do so separately for each element in the list.