lsmeans {emmeans}	R Documentation

Wrappers for alternative naming of EMMs

Description

These are wrappers for emmeans and related functions to provide backward compatibility, or for users who may prefer to use other terminology than “estimated marginal means” – namely “least-squares means” or “predicted marginal means”.

Usage

lsmeans(...)

pmmeans(...)

lstrends(...)

pmtrends(...)

lsmip(...)

pmmip(...)

lsm(...)

pmm(...)

lsmobj(...)

pmmobj(...)

lsm.options(...)

get.lsm.option(x, default = emm_defaults[[x]])

Arguments

...
Arguments passed to the corresponding emxxxx function

x
Character name of desired option

default
default value to return if x not found

Details

For each function with lsxxxx or pmxxxx in its name, the same function named emxxxx is called. Any estimator names or list items beginning with “em” are replaced with “ls” or “pm” before the results are returned

Value

The result of the call to emxxxx, suitably modified.

get.lsm.option and lsm.options remap options from and to corresponding options in the lsmeans options system.

See Also

emmeans, emtrends, emmip, emm, emmobj, emm_options, get_emm_option

Examples

pigs.lm <- lm(log(conc) ~ source + factor(percent), data = pigs)
lsmeans(pigs.lm, "source")

[Package emmeans version 1.4.5 Index]